requirements.txt
sentence-transformers==5.1.0
faiss-cpu==1.11.0
openai==0.28.1
numpy>=1.24
huggingface-hub>=0.16.0
